+++ CHANGELOG.md
@@ -7,7 +7,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
-[Unreleased]: https://github.com/atomist-rugs/rug-editors/compare/0.11.0...HEAD
+[Unreleased]: https://github.com/atomist-rugs/rug-editors/compare/0.12.0...HEAD
+
+## [0.12.0] - 2017-03-01
+
+[0.12.0]: https://github.com/atomist-rugs/rug-editors/compare/0.11.0...0.12.0
+
+README release
+
+### Changed
+
+-   AddTypeScriptEditor and AddTypeScriptGenerator now also add
+    instructions to the project README
 
 ## [0.11.0] - 2017-02-27